+++ src/Map/PollutionCount.cpp
@@ -401,8 +401,8 @@ static void doTask(u16 target_layer, int task_num,
 		if (iVar7 > 0xff)
 			iVar7 = 0xff;
 
-		GXSetTevColor(GX_TEVREG0, (GXColor) { iVar5, iVar5, iVar5, iVar5 });
-		GXSetTevColor(GX_TEVREG1, (GXColor) { iVar7, iVar7, iVar7, iVar7 });
+		GXSetTevColor(GX_TEVREG0, (GXColor) { (u8)iVar5, (u8)iVar5, (u8)iVar5, (u8)iVar5 });
+		GXSetTevColor(GX_TEVREG1, (GXColor) { (u8)iVar7, (u8)iVar7, (u8)iVar7, (u8)iVar7 });
 		s16 x1 = tasks->mX - (tasks->mSize / 2);
 		s16 y1 = tasks->mZ - (tasks->mSize / 2);
 		s16 x2 = tasks->mX + (tasks->mSize / 2);
